A weekend's worth of work creating FFFLL
Both implementations are lacking a function for opening up files. No
other glaring omissions.
The python implementation is fairly complete.
The C implementation currently may-or-may-not segfault depending on
how you initialize the dynamic array and whether or not you add only
stdin and stdout or stderr in addition to those two.