Create standard errmsg functions
errmsg just prints out a message straight to stderr.
errmsgf takes a format string with one %s and sends the formatted
string to errmsg.
errmsgfd takes a format string with one %s and one %d and sends the
formatted string to errmsg.
Fizzbuzz is too slow under valgrind when I'm running all the examples
as a test suite.